Initial CBOR support
For now, just a stand-alone command with basic read support.  Despite
suggestions in the RFC, CBOR isn't really that tight a match for JSON, but many
other JSON libraries also support CBOR, and there may be value in tight
integration to provide CBOR as a serialization option for JSON (JSON to CBOR
and back is lossless, in contrast to the other way around).