Disabled CBOR by default
The anticipated commonalities between JSON and CBOR didn't materialise
and fundamental design conflicts between JSON and CBOR mean it isn't
sensible to tightly couple with this project, and its inclusion was
breaking some platform's build systems.  So disable it by default, with
the future expectation of factoring it out into a separate extension.