Change the suffix rule to an implicit rule, submake jobserver fix
Suffix rules can't have dependencies, but we need the header generated
by the deps rule to build the objects.  Placing the dependency higher
allowed the object builds to fire before the deps build completed.

Remove the hard-coded -j 12 on the tommath sub-make, so that it
coordinates with the main jobserver.